Jacob's commits
Cando's in player and server proxy

Stub for client model and Junit tests

Bug Fixes and Junit tests

Finished as much as I could on JUnit tests for the client model. ...I think the players are supposed to be populated or something, not sure. If not I will add players and test it that way